Crate ci

Runs CI scripts locally on the users computer.

Developers should run this script at least once before pushing to the Runwell repository.

Usage

In order to successfully run this script the user needs to have installed the following programs on their machines:

git

rustup

cargo

hunspell with en_US dictionary

The scripts will eventually install some other Rust or Cargo components:

cargo-miri via rustup

cargo-clippy via rustup

cargo-spellcheck via cargo

Credits

This quality controlling Rust script was heavily inspired by the one used and provided in stdio-utils repository authored by Consolero:

https://github.com/consolero/stdio-utils-rs/blob/dev/0.1/tools/src/bin/quality-control.rs

Developers should run this script before pushing their pull requests to the main repository to make sure that the quality of the pull request’s edits are in the realm of what the project accepts.

Functions

call

Invokes the given command with the provided arguments.

cargo

Invokes the cargo command with the provided arguments.

git

Invokes the git command with the provided arguments.

main

rustup

Invokes the rustup command with the provided arguments.